pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "object_detection_training"
version = "0.0.1"
description = "Object detector training project using rfnet and hydra"
authors = [{name = "Enrique G. Ortiz", email = "ortizeg@gmail.com"}]
dependencies = [
    "hydra-core",
    "loguru",
    "numpy<2.0.0",
    "onnxscript>=0.5.6,<0.6",
    "lightning",
    "pydantic>=2.0",
    "fvcore",
    "wandb>=0.23.1,<0.24",
    "iopath>=0.1.10,<0.2",
    "colored>=2.3.1,<3",
    "rfdetr",
    "torchmetrics",
    "python-dotenv",
    "orjson",
    "tensorboard",
]

[project.scripts]
train-detect = "object_detection_training.task_manager:main"

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable all `E` and `F` codes by default.
# See https://docs.astral.sh/ruff/rules/ for a full list of rules.
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["E203"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect the line length.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["object_detection_training"]

[tool.semantic_release]
version_variable = "pyproject.toml:version"
branch = "main"
build_command = "pip install flit && flit build"
upload_to_pypi = true
upload_to_release = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.develop]
match = "develop"
prerelease = true
prerelease_token = "dev"
